Don't run deploy for every travis build (#52)
* Use travis deployment stage (BETA)

Using a different stage stops the travis deploy command from running for every
build in the build matrix. This could also be accomplished by using a
condition on the deploy to limit it to one build. However, then the deployment
would take place even if other builds fail, while the deploy stage waits for
the previous testing stage to succeed.

Note that this travis feature is **still in beta**.

See also:
https://docs.travis-ci.com/user/build-stages/
https://docs.travis-ci.com/user/build-stages/matrix-expansion/
https://docs.travis-ci.com/user/build-stages/deploy-github-releases/

* Execute travis release stage only when tags present