Revert "is_deleted field should have an index (#85)" (#97)
This reverts commit adfac9be2756115c091d0f28605871179aec3123.

as discussed in https://github.com/Thermondo/thermondo-backend/pull/13998 we actually don't need this any more, and the use-case for other users doesn't exist. 

Heroku Support explanation:

```
Thanks for reaching out.

>Am I safe to assume that Heroku Connect deletes the records in postgres that are trashed in Salesforce? Is the column used only temporarily?

You're correct - we need to clarify this in the documentation. We use the isDeleted field under the hood in sync operations, but as soon as a record is soft-deleted in Salesforce, we hard-delete it from Postgres. So deleted records will get removed from the Postgres table altogether.

>If yes, what's the time between the deleted record appearing and it being deleted?

It would happen in the normal sync process, so, within 2 minutes if using normal polling, or as the change is made if using accelerating polling/streaming API.

Hope that helps, please let us know is we can assist further.
```